// src/common/constant/api-error-code.enum.ts
export enum ApiErrorCode {
  PERMISSION_EXISTED = 10001,
  ROLE_EXISTED = 10002,
  INVALID_CAPTCHA = 10003,
  CAPTCHA_ERROR = 10004,
  USER_EXISTED = 10005,
  USER_UN_EXISTED = 10006,
  PASSWORD_ERROR = 10007,
  USER_FREEZED = 10008,
  USER_UNFREEZE = 10009,
  EMPTY_TOKEN = 10010,
  INVALID_TOKEN = 10011,
}